requirements.txt
streamlit
pillow
numpy
piexif
